requirements.txt
pytube3==9.6.4
tqdm==4.46.1
typing-extensions==3.7.4.2
